ecosense-backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
tensorflow-cpu==2.12.0
numpy>=1.22.0
Pillow==9.0.0
#firebase-admin
asyncio
asgiref
opencv-python==4.8.0.74
google-cloud-firestore==2.3.4